# Optimize ToByteArray()