Update convenience method ToByteArray() to use the new API.

Sample benchmark serializing a tiny protobuf message,
|                        Method |     Mean |    Error |   StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|------------------------------ |---------:|---------:|---------:|-------:|------:|------:|----------:|
| ToByteArray_CodedOutputStream | 76.82 ns | 1.079 ns | 1.009 ns | 0.0279 |     - |     - |      88 B |
|           ToByteArray_WriteTo | 53.73 ns | 0.833 ns | 0.779 ns | 0.0076 |     - |     - |      24 B |