@jtattermusch 2021-08-16
FTR, Can you link to the implementation you're referring to as "new API"?