@jtattermusch 2021-08-16
> FTR, Can you link to the implementation you're referring to as "new API"?

Is it this one?
https://github.com/protocolbuffers/protobuf/blob/98b6ead7756237cfbad9006c6032d96dbc82fae3/csharp/src/Google.Protobuf/MessageExtensions.cs#L182