@xqrzd 2021-08-16
Yep, basically this PR makes it so if you were already doing,
```csharp
var size = messagePb.CalculateSize();
var buffer = new byte[size];
messagePb.WriteTo(buffer);
```
you can now do,
```csharp
var buffer = messagePb.ToByteArray();
```
and get the same performance.