@jtattermusch 2021-08-19
Tests are actually failing.
```
A total of 1 test files matched the specified pattern.
  Failed IntermixingOfNewAndLegacyGeneratedCodeWorksWithCodedInputStream [19 ms]
  Error Message:
   Google.Protobuf.InvalidProtocolBufferException : Message LegacyGeneratedCodeMessageA doesn't provide the generated method that enables WriteContext-based serialization. You might need to regenerate the generated protobuf code.
  Stack Trace:
     at Google.Protobuf.WritingPrimitivesMessages.WriteRawMessage(WriteContext& ctx, IMessage message) in /tmp/protobuf/protobuf/csharp/src/Google.Protobuf/WritingPrimitivesMessages.cs:line 96
   at Google.Protobuf.LegacyGeneratedCodeTest.ParseContextEnabledMessageB.pb::Google.Protobuf.IBufferMessage.InternalWriteTo(WriteContext& output) in /tmp/protobuf/protobuf/csharp/src/Google.Protobuf.Test/LegacyGeneratedCodeTest.cs:line 242
   at Google.Protobuf.MessageExtensions.WriteTo(IMessage message, Span`1 output) in /tmp/protobuf/protobuf/csharp/src/Google.Protobuf/MessageExtensions.cs:line 186
   at Google.Protobuf.MessageExtensions.ToByteArray(IMessage message) in /tmp/protobuf/protobuf/csharp/src/Google.Protobuf/MessageExtensions.cs:line 113
   at Google.Protobuf.LegacyGeneratedCodeTest.IntermixingOfNewAndLegacyGeneratedCodeWorksWithCodedInputStream() in /tmp/protobuf/protobuf/csharp/src/Google.Protobuf.Test/LegacyGeneratedCodeTest.cs:line 49
  Failed LegacyGeneratedCodeThrowsWithReadOnlySequence [1 ms]
  Error Message:
   Google.Protobuf.InvalidProtocolBufferException : Message LegacyGeneratedCodeMessageA doesn't provide the generated method that enables WriteContext-based serialization. You might need to regenerate the generated protobuf code.
  Stack Trace:
     at Google.Protobuf.WritingPrimitivesMessages.WriteRawMessage(WriteContext& ctx, IMessage message) in /tmp/protobuf/protobuf/csharp/src/Google.Protobuf/WritingPrimitivesMessages.cs:line 96
   at Google.Protobuf.LegacyGeneratedCodeTest.ParseContextEnabledMessageB.pb::Google.Protobuf.IBufferMessage.InternalWriteTo(WriteContext& output) in /tmp/protobuf/protobuf/csharp/src/Google.Protobuf.Test/LegacyGeneratedCodeTest.cs:line 242
   at Google.Protobuf.MessageExtensions.WriteTo(IMessage message, Span`1 output) in /tmp/protobuf/protobuf/csharp/src/Google.Protobuf/MessageExtensions.cs:line 186
   at Google.Protobuf.MessageExtensions.ToByteArray(IMessage message) in /tmp/protobuf/protobuf/csharp/src/Google.Protobuf/MessageExtensions.cs:line 113
   at Google.Protobuf.LegacyGeneratedCodeTest.LegacyGeneratedCodeThrowsWithReadOnlySequence() in /tmp/protobuf/protobuf/csharp/src/Google.Protobuf.Test/LegacyGeneratedCodeTest.cs:line 74

Failed!  - Failed:     2, Passed:  1140, Skipped:     0, Total:  1142, Duration: 7 s - /tmp/protobuf/protobuf/csharp/src/Google.Protobuf.Test/bin/Release/netcoreapp2.1/Google.Protobuf.Test.dll (netcoreapp2.1)
```
https://source.cloud.google.com/results/invocations/f9718809-b44b-43db-bb96-22b8a23370ba/log

I think the issue is that the Span-based codepath will not exist for messages generated by old codegen (which is what the test is complaining about) - but the library still needs to stay compatible with such messages (especially since ToByteArray() method has been in the public API basically since the library's inception).
In order to make this work, you'd need to intelligently detect if using the `WriteTo(this IMessage message, Span<byte> output)` method is safe for given message.