@xqrzd 2021-08-19
Ah I see. For intelligently checking I came up with this,
```csharp
[SecuritySafeCritical]
public static byte[] ToByteArray(this IMessage message)
{
    ProtoPreconditions.CheckNotNull(message, "message");
    byte[] result = new byte[message.CalculateSize()];

    if (message is IBufferMessage)
    {
        message.WriteTo(result);
    }
    else
    {
        CodedOutputStream output = new CodedOutputStream(result);
        message.WriteTo(output);
        output.CheckNoSpaceLeft();
    }

    return result;
}
```

which gets the failures down to 2. Unfortunately, they're mixed cases, which would require checking more than just the argument. I don't think it's worth it, given how minor this is, and that there's an alternative that achieves the same performance.

Maybe a new extension method could be added that does,
```csharp
var size = messagePb.CalculateSize();
var buffer = new byte[size];
messagePb.WriteTo(buffer);
```